Support newer versions of gym
Version 0.9.6 of gym introduced several changes which broke
compatibility with most environments, including updates to the
MultiDiscrete interface and changes to the step, render, and observe
methods from private to public. This patch aims to add compatibility
with newer versions of gym by addressing the following:

* Convert MultiDiscrete action spaces to 3D-Box spaces to allow integers
to be sampled from a set lower and upper-bound.
* Change all private methods required by gym to public methods (ie.
change `_step()` to `step()`).

Signed-Off-By: Robert Clark <robdclark@outlook.com>